refactor: rename retry class
