+++ src/uploadx/lib/retry-handler.spec.ts
@@ -1,8 +1,8 @@
-import { ErrorHandler, ErrorType } from './error-handler';
+import { ErrorType, RetryHandler } from './retry-handler';
 
 describe('ErrorHandler', () => {
   it('ErrorHandler.kind(status)', () => {
-    const errorHandler = new ErrorHandler();
+    const errorHandler = new RetryHandler();
     expect(errorHandler.kind(400)).toBe(ErrorType.Fatal);
     expect(errorHandler.kind(0)).toBe(ErrorType.Retryable);
     expect(errorHandler.kind(500)).toBe(ErrorType.Retryable);
